Move Profile data class under java/data/user folder
