+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/profile/edit/ProfileEditContract.kt
@@ -1,7 +1,7 @@
 package com.bounswe2020group3.paperlayer.profile.edit
 
 import com.bounswe2020group3.paperlayer.mvp.Mvp
-import com.bounswe2020group3.paperlayer.profile.data.Profile
+import com.bounswe2020group3.paperlayer.data.user.Profile
 import com.bounswe2020group3.paperlayer.data.user.User
 
 interface ProfileEditContract {
